Move remaining configuration into convention plugin
